Refactor slow-running benchmark to slow_benchmarks directory
